polki_dolki/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.11.1
certifi==2022.6.15
Django==3.2.5
django-ajax-selects==2.2.0
django-bootstrap4==3.0.1
django-datatables-view==1.20.0
django-filter==22.1
django-guardian==2.4.0
django-material==1.5.8
django-materializecss-form==1.1.17
django-utils-six==2.0
djangorestframework==3.13.1
importlib-metadata==2.0.0
Markdown==3.3.4
mysqlclient==2.1.1
pip==21.2.4
psycopg2==2.8.6
Pygments==2.12.0
pytz==2022.1
setuptools==61.2.0
soupsieve==2.3.1
sqlparse==0.4.1
typing_extensions==4.1.1
wheel==0.37.1
wincertstore==0.2
zipp==3.8.0
